Adds an intermediary booking confirmation window for the traveler to see their booking details and the booking's total cost prior to submitting their booking.
